requirements.txt
torch
fair-esm
biopython
pandas
numpy
scikit-learn
lightgbm
matplotlib
kaggle
